dacycler: make additive-inflation Z dtype-relative (fp32 4D cycling)
ETKF._apply_additive drew its structured-perturbation recombination matrix
with jax.random.normal(key, shape) and no dtype. dabench force-enables
jax_enable_x64 at import, so jax.random.normal then defaults to float64,
promoting a float32 analysis to float64 (E = Xb_pert @ Z) and breaking the
per-window jax.lax.scan carry in DACycler.cycle ("carry input and output must
have equal types") whenever additive_inflation > 0 -- the second fp32-4D
blocker (the first was the model-forecast promotion, fixed on the MLTLM side).

Fix: pin Z to Xa_pert.dtype, honouring the dtype invariant (derive every
intermediate from the input array dtype, never force one). fp64 in -> float64
Z (byte-identical; full ETKF4D/LETKF4D/cross-comparison suites unchanged);
fp32 in -> float32 Z, so the analysis stays fp32 and the scan carry matches.

+++ dabench/dacycler/_etkf.py
@@ -438,7 +438,13 @@ class ETKF(dacycler.DACycler):
         if sigma <= 0.0:
             return Xa_pert
         ensemble_dim = Xa_pert.shape[1]
-        Z = jax.random.normal(key, (ensemble_dim, ensemble_dim))
+        # dtype-relative Z: jax.random.normal defaults to float64 when
+        # jax_enable_x64 is on (dabench force-enables it), which would promote a
+        # float32 analysis to float64 (E = Xb_pert @ Z) and break the fp32
+        # lax.scan carry.  Pin Z to the perturbation dtype (the DTYPE INVARIANT:
+        # derive every intermediate from the input array dtype, never force one).
+        Z = jax.random.normal(key, (ensemble_dim, ensemble_dim),
+                              dtype=Xa_pert.dtype)
         E = Xb_pert @ Z
         E = E - jnp.mean(E, axis=1, keepdims=True)   # mean-zero across members
         rms = jnp.sqrt(jnp.mean(E ** 2))
